+++ library/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Add `getDefault` method to get default value of schema (issue #105)
 - Add support for enums as key of `record` and `recordAsync` schema (issue #134)
 - Add support for default values to `optional`, `optionalAsync`, `nullable`, `nullableAsync`, `nullish` and `nullishAsync` schema (issue #96, #118)
 - Deprecate `withDefault` method in favor of `optional` schema